requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1637088766493/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1649013200726/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1636093929600/work
appnope==0.1.3
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
backcall==0.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
blinker==1.4
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854217904/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1640686991047/work
certifi==2022.6.15
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1636046166270/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
click @ file:///Users/runner/miniforge3/conda-bld/click_1651215461185/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1653061851209/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1649035278201/work
cycler==0.11.0
Cython==0.29.30
cytoolz==0.11.2
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1651516556145/work
decorator==5.1.1
et-xmlfile==1.1.0
executing==0.8.3
ffmpeg==1.4
ffpyplayer==4.3.5
filelock==3.7.0
flatbuffers==2.0
flopco-pytorch==0.1.4
flopth==0.0.502
fonttools==4.33.3
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1648772000562/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1653010523205/work
gast==0.5.3
gdown==4.4.0
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1650592442422/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
graphviz==0.20
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1648220970454/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964070496/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imagecodecs @ file:///Users/runner/miniforge3/conda-bld/imagecodecs_1652925992212/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1652789504106/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1648728616176/work
imutils==0.5.4
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1651821407140/work
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1648882471605/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1653162972907/work
numpy==1.22.4
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5
openpyxl==3.0.9
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.2
parso==0.8.3
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1617910651905/work
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1649818082882/work
ply==3.11
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.12.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1638819640841/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1648857342405/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pythran==0.11.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets @ file:///Users/runner/miniforge3/conda-bld/pywavelets_1649616478789/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1648757198437/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1637781155505/work
scikit-image @ file:///Users/runner/miniforge3/conda-bld/scikit-image_1645196796820/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1652976912267/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1653074167280/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
stack-data==0.2.0
tabulate==0.8.9
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1651519171059/work/tensorboard-2.9.0-py3-none-any.whl
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1645578792360/work
tensorflow-docs @ git+https://github.com/tensorflow/docs@3550667e06ea24580b6d907aaf09f0c8e0dfca23
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1651666912743/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1636177962100/work
torch @ file:///Users/runner/miniforge3/conda-bld/pytorch-recipe_1647804319176/work
torchvision @ file:///Users/runner/miniforge3/conda-bld/torchvision-split_1628627505561/work
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1650370875435/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
wcwidth==0.2.5
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
wrapt==1.14.1
xmltodict==0.13.0
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1648966573015/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
